mingle-requirements.txt
cachetools==3.1.1
certifi==2021.10.8
chardet==4.0.0
django-ckeditor==6.3.2
django-ckeditor-5==0.1.8
django-js-asset==1.2.3
enum34==1.1.10
google-auth==1.35.0
googleapis-common-protos==1.52.0
idna==2.10
packaging==20.9
protobuf==3.17.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.8
pytz==2024.1
qrcode==6.1
razorpay==1.4.2
requests==2.27.1
rsa==4.5
six==1.16.0
twilio==6.63.2
urllib3==1.26.16
